+++ cup/net/async/common.py
@@ -20,44 +20,39 @@ __all__ = [
 
 def ip_port2connaddr(peer):
     """
-    connaddr是个64bit的int
+    connaddr is a 64bit int
         32 -  16    - 16   - 32
         ip - port   - stub - future
-    该函数以peer(ipaddr, port)为输入参数， 生成一个connaddr
+
+    :param peer:
+        (ipaddr, port)
+    :return:
+        return a connaddr
     """
     misc.check_type(peer, tuple)
     ipaddr, port = peer
     misc.check_type(ipaddr, str)
     packed = socket.inet_aton(ipaddr)
-    return ((struct.unpack("!L", packed)[0] << 64) | (port << 48))
+    return (struct.unpack("!L", packed)[0] << 64) | (port << 48)
 
 
 def add_stub2connaddr(pack, stub):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数设置connaddr中的stub到connaddr
+    add stub into connaddr
     """
-    return (pack | (stub << 32))
+    return pack | (stub << 32)
 
 
 def add_future2connaddr(pack, future):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数设置connaddr中的future到connaddr
+    add future into connaddr
     """
-    return (pack | future)
+    return pack | future
 
 
 def get_ip_and_port_connaddr(pack):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数获取connaddr中的(ip, port)
+    get (ip, port) from connaddr
     """
     ipaddr = getip_connaddr(pack)
     port = getport_connaddr(pack)
@@ -66,41 +61,29 @@ def get_ip_and_port_connaddr(pack):
 
 def getip_connaddr(pack):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数获取connaddr中的ip
+    get ip from connaddr
     """
     return socket.inet_ntoa(struct.pack('!L', pack >> 64))
 
 
 def getport_connaddr(pack):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数获取connaddr中的port
+    get port from connaddr
     """
-    return ((pack >> 48) & 0xffff)
+    return (pack >> 48) & 0xffff
 
 
 def getstub_connaddr(pack):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数获取connaddr中的stub
+    get stub from connaddr
     """
-    return ((pack >> 32) & 0xffff)
+    return (pack >> 32) & 0xffff
 
 
 def getfuture_connaddr(pack):
     """
-    connaddr是个64bit的int
-        32 -  16    - 16   - 32
-        ip - port   - stub - future
-    该函数获取connaddr中的future
+    get future from conaddr
     """
-    return ((pack) & 0xffff)
+    return (pack) & 0xffff
 
 # vi:set tw=0 ts=4 sw=4 nowrap fdm=indent
